requirements.txt
pillow==11.3.0
openai==1.101.0
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20